[codex] RX ring境界の追い越し判定を修正

+++ src/lidar/parser/device_info.c
@@ -11,13 +11,13 @@ bool read_device_info_frame(ParserDeviceInfo* info)
                 return false;
 
         // Manual section 3.3: firmware low byte is major, high byte is minor.
-        info->model            = (uint8_t)read_byte();
-        info->firmware_major   = (uint8_t)read_byte();
-        info->firmware_minor   = (uint8_t)read_byte();
-        info->hardware_version = (uint8_t)read_byte();
+        info->model            = read_byte();
+        info->firmware_major   = read_byte();
+        info->firmware_minor   = read_byte();
+        info->hardware_version = read_byte();
         // Preserve the serial bytes in wire order; no integer endian conversion.
         for (size_t i = 0; i < SYS_PACKET_DEVICE_SERIAL_SIZE; ++i)
-                info->serial_number[i] = (uint8_t)read_byte();
+                info->serial_number[i] = read_byte();
 
         return true;
 };
